fix: update models api
