+++ src/state/modelStore.ts
@@ -1,0 +1,22 @@
+import { create } from 'zustand'
+
+interface Model {
+  id: string
+  name: string
+}
+type Store = {
+  models: Model[]
+  newModel: Model
+  addModel: () => void
+}
+
+const useModelStore = create<Store>((set) => ({
+  models: [],
+  newModel: '',
+  addModel() {
+    set((state) => ({
+      ...state,
+      models: addModel(state.models, this.newModel),
+    }))
+  },
+}))
