+++ src/state/trainingRequestStore.ts
@@ -1,0 +1,17 @@
+import { create } from 'zustand'
+
+
+interface TrainingRequestState {
+    id: string,
+
+}
+
+type Store = {
+    trainingRequest: TrainingRequestState
+    addTrainingRequest : ()=> void
+
+}
+
+export const useTrainingRequestStore = create<TrainingRequestState>(
+  
+)
